Convert Kubernetes launcher to a scheduler
This launcher listened only for services, and in response created
sources.

This scheduler combines bits of the AD scheduler (the bits listening for
service configs and translating them to services) and bits of the
launcher (translating services to sources), with minimal changes to
either.

The code duplication is in code that will soon be refactored, so I hope
you'll excuse it for now.