Handle integration.Config's in the run() goroutine
This minor refactor switches from translating integration.Config ->
service.Service in the callback from the MetaScheduler, to performing
that translation in the run() goroutine.  It also moves the functions in
the source file so that Config -> Service and Service -> Source occur in
that order.  This should make diffs for subsequent refactorings easier
to read.