[tagger] Remove GetEntityHash
This functionality is already provided by TagWithHash. Its only user are
Autodiscovery Services, and `service.GetTags` is already supposed to
return a hash. This way, we reduce the interface of the global tagger,
and avoid calls to the tagger for services that don't have an entity
there to begin with (as in, non container based services).