Introduce a new priorityConfigManager
This config manager is disabled by default, and enabled by the temporary
`logs_config.cca_use_bare_configs` config flag.  Without that flag, the
existing simpleConfigManager is used.

Currently, the priorityConfigManager does exactly the same thing as the
simpleConfigManager, and the tests demonstrate that.  But it does so in
a way that focuses on generation of configs on a per-service basis.
This commit should be reviewed as a re-implementation of the existing
algorithm, setting the stage for a simple change in the next commit.