Implement TemplatePriority
This commit makes a simple change to the config manager which has no
effect when all templates have TemplatePriority 0 (confirmed by
continuing to pass all existing tests), but adds the feature that a
service only schedules templates in the highest priority band available.

This functionality will be used in subsequent commits.