Add telemetry about container event bundling
The docker and kubernetesapiserver checks bundle events coming from the
Docker daemon and API Server, respectively, based on a few rules. We
want to monitor how that bundling works, and estimate the impact of a
possible unbundling of those events, so we are adding telemetry to that
effect.

This commit adds the following metrics:

* kubernetes_apiserver_kube_events: kubernetes events processed by the
  check
* kubernetes_apiserver_emitted_events: number of bundles emitted as
  datadog events
* docker_events: docker events processed by the check
* docker_emitted_events: number of bundles emitted as datadog events