fix verifier error
Caluclating read_len earlier in the function and returning it later
causes the compiler to save the value on to the stack, causing the
verifier on older kernels to loose the state information.
http->fragment_len is then set by offseting into the stack and setting
it to read_len.
This causes the verifier to complain about arithmetic between fp and a
register with an unbounded value.
Directly returning the read_len keeps the value in the register, and
preserves the state information.